rearrange: create cat module
